==> R/0_setup.R <==
devtools::install_github('chris-mcginnis-ucsf/MULTI-seq')

dir.create("R")
dir.create("data")
dir.create("results")
dir.create("plots")
dir.create(file.path("plots","umaps"))
dir.create(file.path("plots","heatmaps"))
dir.create(file.path("plots","others"))
